Confirmations is now uint32

+++ JSON.go
@@ -293,7 +293,7 @@ type RawTransaction struct {
 	Vin           []Vin  `json:"vin"`
 	Vout          []Vout `json:"vout"`
 	BlockHash     string `json:"blockhash,omitempty"`
-	Confirmations int64  `json:"confirmations,omitempty"`
+	Confirmations uint32 `json:"confirmations,omitempty"`
 	Time          int64  `json:"time,omitempty"`
 	Blocktime     int64  `json:"blocktime,omitempty"`
 }
